pyproject.toml
[project]
name = "hermes-assistant"
version = "0.1.0"
description = "A simple chat agent built using the Hermes Agent Python library"
requires-python = ">=3.11"
dependencies = [
    "hermes-agent @ git+https://github.com/NousResearch/hermes-agent.git",
    "python-dotenv>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "gradio>=5.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
chatbot = "chatbot:main"
